requirements.txt
# Core Dependencies
Flask>=2.3.0
Flask-CORS>=4.0.0
SQLAlchemy>=2.0.0
bcrypt>=4.0.0
PyJWT>=2.8.0
python-dateutil>=2.8.0
requests>=2.31.0
psutil>=5.9.0

# Network & Security
scapy>=2.5.0
netaddr>=0.8.0
cryptography>=41.0.0

# Background Tasks
celery>=5.3.0
redis>=4.6.0

# Web Interface
Jinja2>=3.1.0
Werkzeug>=2.3.0

# Monitoring & Logging
prometheus-client>=0.17.0

# Development Dependencies (optional)
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0

# Documentation (optional)
sphinx>=7.1.0
sphinx-rtd-theme>=1.3